🚚 api: move CertificationCandidateSubscription read-model to src/shared
